Refactor v1-specific API calls for uploading
- Create separate subpackages for the v1 and v2 clients for Keeper.
- Refactor the upload command handler so that the upload itself is done
  through an all-inclusive service function to make it easier to have v1
  and v2 implementations.
